requirements.txt
Flask>=2.0
google-api-python-client>=2.0
google-auth-oauthlib>=0.5
google-auth-httplib2>=0.1
requests # Often needed by google libraries
